requirements.txt
pydantic>=2.9.0
pydantic-ai>=0.0.14
python-dotenv>=1.0.1
httpx>=0.27.0
python-dotenv